Crate gcm

gcm

Usage:

Add this to Cargo.toml:

[dependencies]
gcm = "0.1.3"

then add this to your crate root:

extern crate gcm;

Examples:

Here is an example to send out a GCM Message with some custom data:

use gcm::Message;
use std::collections::HashMap;

let mut map = HashMap::new();
map.insert("message", "Howdy!");
 
let result = Message::new("<registration id>")
    .data(map)
    .send("<GCM API Key>");

To send a message using GCM Notifications, we first build the notification:

use gcm::{Message, NotificationBuilder};

let notification = NotificationBuilder::new("Hey!")
    .body("Do you want to catch up later?")
    .finalize();

And then set it in the message, before sending it:

let result = Message::new("<registration id>")
    .notification(notification)
    .send("<GCM API Key>");

You can now handle the result accordingly:

match result {
  Ok(response) => println!("message_id: {:?}", response.message_id),
  Err(error) => println!("Error: {:?}", error),
}

Reexports

pub use message::response::*;

Modules

response

Structs

Message
Represents a GCM message. Construct the GCM message using various utility methods and finally send it.

Examples:

use gcm::Message;

Notification
This struct represents a GCM notification. Use the corresponding NotificationBuilder to get an instance. You can then use this notification instance when sending a GCM message.

NotificationBuilder
A builder to get a Notification instance.

Enums

Error

Priority